RestAPI/requirements_api.txt
pandas
requests
flask
cassandra-driver
